setup onItemListener move adapter to it's own package
